Source both sqm's functions.sh and defaults.sh to enable erro logging
The sqm_debug, sqm_error and sqm_trace functions require us to
source both functions.sh and defaults.sh.
I am not sure whether that change is actually desired, but
it will make the error/info reporting in the new locking
code actually work as intended. But this is just a proposal
in code...

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>